pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "alx_connect"
version = "1.0.0"
description = "ALX Connect Hub - Mentor matching and interview coaching platform"
readme = "README.md"
requires-python = ">=3.8"
license = {text = "MIT"}
authors = [
    {name = "ALX Connect Team", email = "team@alxconnect.com"}
]
dependencies = [
    "fastapi>=0.68.0",
    "uvicorn>=0.15.0",
    "python-dotenv>=0.19.0",
    "pydantic>=2.0.0",
    "numpy>=1.21.0",
    "pandas>=1.3.0",
    "scikit-learn>=1.0.0",
    "langchain>=0.1.0",
    "langchain-groq>=0.1.0",
    "langchain-community>=0.0.10",
    "langchain-huggingface>=0.0.5",
    "faiss-cpu>=1.7.0",
    "groq>=0.1.0",
    "llama-parse>=0.1.0",
    "nest-asyncio>=1.5.0",
    "python-multipart>=0.0.5",
]

[project.optional-dependencies]
dev = [
    "black>=22.0.0",
    "isort>=5.10.0",
    "flake8>=4.0.0",
    "pytest>=7.0.0",
    "pytest-asyncio>=0.18.0",
    "pytest-cov>=4.0.0",
]

[tool.black]
line-length = 88
target-version = ['py38']
include = '\.pyi?$'

[tool.isort]
profile = "black"
multi_line_output = 3
line_length = 88

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "-v --cov=alx_connect --cov-report=term-missing" 